code refactor

+++ game/__init__.py
@@ -137,12 +137,12 @@ class Game:
         )
 
     async def stop(self):
+        if self.is_stopped: return
         print("======= Game stopped =======")
         self.is_stopped = True
         self.next_flag.clear()
         await self.cancel_running_task(self.task_game_loop)
         await self.cancel_running_task(self.task_run_timer_phase)
-
 
         if self.players:
             await self.delete_channel()
